修正代码，timeout 也转换类型

+++ src/Redis/RedisManager.php
@@ -190,7 +190,7 @@ class RedisManager
             }
             else
             {
-                $redis->connect($host, (int)$config['port'] ?? 6379, $config['timeout'] ?? 0);
+                $redis->connect($host, (int) ($config['port'] ?? 6379), (float) ($config['timeout'] ?? 0));
             }
             if (('' !== ($config['password'] ?? '')) && !$redis->auth($config['password']))
             {
